+++ src/app/app-routing.module.ts
@@ -37,10 +37,25 @@ const routes: Routes = [
                     ),
             },
             {
+                path: 'inference',
+                canActivate: [AuthenticationGuard],
+                loadChildren: () =>
+                    import('@app/modules/inference/inference.module').then(
+                        (m) => m.InferenceModule
+                    ),
+            },
+            {
                 path: 'forbidden',
                 loadChildren: () =>
                     import('@modules/forbidden/forbidden.module').then(
                         (m) => m.ForbiddenModule
+                    ),
+            },
+            {
+                path: 'profile',
+                loadChildren: () =>
+                    import('@modules/profile/profile.module').then(
+                        (m) => m.ProfileModule
                     ),
             },
             {